backend/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
chainlit>=1.7
httpx>=0.25
pydantic>=1.10
python-dotenv>=1.0
google-generativeai>=0.3.0
langchain>=0.1
